+++ example/lib/widgets/_widgets.dart
@@ -4,4 +4,5 @@ export 'example_dialog.dart';
 export 'example_dialog_no_buttons.dart';
 export 'example_dialog_non_dismissible.dart';
 export 'example_scrollable_bottom_sheet.dart';
+export 'yg_debug_bottom_sheet.dart';
 export 'yg_screen.dart';
